bug fix and refactoring physics

+++ Engine/src/main/java/components/GridLines.java
@@ -16,20 +16,20 @@ public class GridLines extends Component {
         Vector2f CameraPos = camera.position;
         Vector2f projectionSize = camera.getProjectionSize();
 
-        int firstX = ((int) (CameraPos.x / Settings.GRID_WIDTH)) * Settings.GRID_WIDTH;
-        int firstY = ((int) (CameraPos.y / Settings.GRID_HEIGHT))* Settings.GRID_HEIGHT;
+        float firstX = ((int) (CameraPos.x / Settings.GRID_WIDTH)) * Settings.GRID_WIDTH;
+        float  firstY = ((int) (CameraPos.y / Settings.GRID_HEIGHT))* Settings.GRID_HEIGHT;
 
-        int verticalLines = (int) (projectionSize.x*camera.getZoom() / Settings.GRID_WIDTH);
-        int horizontalLines = (int) (projectionSize.y*camera.getZoom() / Settings.GRID_HEIGHT);
+        float  verticalLines = (int) (projectionSize.x*camera.getZoom() / Settings.GRID_WIDTH);
+        float  horizontalLines = (int) (projectionSize.y*camera.getZoom() / Settings.GRID_HEIGHT);
 
-        int height = (int) (projectionSize.y*camera.getZoom())+Settings.GRID_WIDTH;
-        int width = (int)( projectionSize.x*camera.getZoom())+Settings.GRID_WIDTH;
+        float  height = (int) (projectionSize.y*camera.getZoom())+Settings.GRID_WIDTH;
+        float  width = (int)( projectionSize.x*camera.getZoom())+Settings.GRID_WIDTH;
 
-        int maxLines=Math.max(verticalLines,horizontalLines);
+        float  maxLines=Math.max(verticalLines,horizontalLines);
         Vector3f color=new Vector3f(0.2f,0.2f,0.2f);
         for(int i=0;i<maxLines;i++){
-            int x=firstX+(Settings.GRID_WIDTH*i);
-            int y=firstY+(Settings.GRID_HEIGHT*i);
+            float  x=firstX+(Settings.GRID_WIDTH*i);
+            float  y=firstY+(Settings.GRID_HEIGHT*i);
             if(i<verticalLines){
                 DebugDraw.addLine2D(new Vector2f(x,firstY),new Vector2f(x,firstY+height),color);
